fix:moved getClientContext and connectAccount methods to utils file #170
