+++ packages/client-ts/src/utils.ts
@@ -1,8 +1,10 @@
 import EncryptionUtils from '@verida/encryption-utils';
+import { AutoAccount } from '@verida/account-node';
+import { Client, Context, EnvironmentType, Network } from './index';
+import config from './config';
 import url from 'url';
 import { PermissionOptionsEnum } from './context/interfaces';
 import { FetchUriParams } from './interfaces';
-import Client from './client';
 const bs58 = require('bs58');
 
 /**
@@ -80,7 +82,7 @@ export function explodeVeridaUri(uri: string): FetchUriParams {
  * @param context An existing context used to open the external database
  * @returns
  */
-export async function fetchVeridaUri(uri: string, context: any): Promise<string> {
+export async function fetchVeridaUri(uri: string, context: any): Promise<object> {
 	const url = explodeVeridaUri(uri);
 
 	const db = await context.openExternalDatabase(url.dbName, url.did, {
@@ -111,4 +113,53 @@ export async function fetchVeridaUri(uri: string, context: any): Promise<string>
 
 		throw err;
 	}
+}
+
+
+
+export async function connectAccount(privateKey: string, contextName: string, environment: EnvironmentType): Promise<Context> {
+
+	const context = await Network.connect({
+		context: {
+			name: contextName,
+		},
+		client: {
+			environment: environment,
+		},
+		account: new AutoAccount(
+			{
+				defaultDatabaseServer: {
+					type: 'VeridaDatabase',
+					endpointUri: config.environments[environment].defaultDatabaseServerUrl as string,
+				},
+				defaultMessageServer: {
+					type: 'VeridaMessage',
+					endpointUri: config.environments[environment].defaultMessageServerUrl as string,
+				},
+			},
+			{
+				privateKey: privateKey,
+				environment: EnvironmentType.TESTNET,
+			}
+		),
+	});
+	return context as Context;
+};
+
+
+export async function getClientContext(uri: string, environment: EnvironmentType): Promise<Context> {
+
+	const clientConfig = {
+		environment: environment,
+		didServerUrl: config.environments[environment].didServerUrl
+	}
+
+	const url = explodeVeridaUri(uri)
+
+	const context = await new Client(clientConfig).openExternalContext(
+		url.contextName,
+		url.did
+	);
+
+	return context
 }
