+++ packages/verifiable-credentials/ve14e08606e47aaa6c53766fa3a6315d2d131d392583ab1203a7284fbba25687f/CURRENT
@@ -1,0 +1,1 @@
+MANIFEST-000002
